backend/requirements.txt
Django>=4.2
djangorestframework>=3.14
drf-spectacular>=0.27
pytest>=8.0
pytest-django>=4.5